requirements.txt
tqdm==4.66.1
librosa==0.10.1
hydra-core==1.3.2
joblib==1.3.2
matplotlib==3.4.2
numba==0.53.1
scikit-learn==0.24.2
scipy==1.6.3
sklearn==0.0
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
wget==3.2
zipp==3.4.1
efficientnet-lite-pytorch==0.1.0
efficientnet-pytorch==0.7.0
pycm==4.0
